material: Add material parameter
Matparam (or Material parameter) has two goals, first to simplify the creation of
new materials/shaders by allowing to automatically set a shader uniforms values.
The second goal is to be able to load material definition from an asset file, this mean
to create a new material at runtime.

Finally the render function can in the future able to do some rendering optimization
with knowing the different used materials and their params.